Put more of the functionality in the base module
We can reduce compile sizes by not generating code where it's not
needed. Additionally, this allows the documentation for those functions
to be more easily exposed (because you can see them in the docs for this
package and don't have to build docs for your application to see it).

Finally, we also add an optional `timeout` to the `call` and `call!`
functions so that we preserve the ability to use custom timeouts when
making `GenServer.call/3` invocations.