Support `wait_max` value of zero, meaning don't wait
If someone does *not* want to wait for a remote process to start,
perhaps because they have other retry mechanisms that will kick in if
they can't get a process started in time, they should be able to set
`wait_max` on the Horde Process to zero and bypass that entire
functionality.

If someone does this, then `:ignore` is returned immediately if
`start/1` returns `:ignore`.